+++ openzen/examples/ExamplePython.py
@@ -2,12 +2,23 @@
 #
 # OpenZen Python example
 #
-# Make sure the openzen.pyd (for Windows) or openzen.so (Linux/Mac)
-# are in the same folder as this file.
+# Make sure the openzen.pyd (for Windows) or openzen.so (Linux/Mac, simply rename 
+# libOpenZen.so to openzen.so) are in the same folder as this file.
+#
 # If you want to connect to USB sensors on Windows, the file SiUSBXp.dll
 # should also be in the same folder.
 #
+# Python interfaces definitions could be find in `../src/bindings/OpenZenPython.cpp`.
+# Function, enum and property names of the OpenZen Python interface are in double quotes.
+#
 ###########################################################################
+
+
+#
+# Python examples on connecting to sensor, toggling sensor settings and print sensor data.
+#
+# Check our docs for more https://lpresearch.bitbucket.io/openzen/latest/getting_started.html
+#
 
 import sys
 import openzen
@@ -16,7 +27,7 @@ openzen.set_log_level(openzen.ZenLogLevel.Warning)
 
 error, client = openzen.make_client()
 if not error == openzen.ZenError.NoError:
-    print ("Error while initializinng OpenZen library")
+    print ("Error while initializing OpenZen library")
     sys.exit(1)
 
 error = client.list_sensors_async()
@@ -43,11 +54,17 @@ if sensor_desc_connect is None:
     print("No sensors found")
     sys.exit(1)
 
-# connect to the first sensor found
+# connect to the first sensor found, more on https://lpresearch.bitbucket.io/openzen/latest/io_systems.html
 error, sensor = client.obtain_sensor(sensor_desc_connect)
 
 # or connect to a sensor by name
-#error, sensor = client.obtain_sensor_by_name("LinuxDevice", "LPMSCU2000003")
+# error, sensor = client.obtain_sensor_by_name("SiUsb", "ig1pcan000028", 921600)
+
+# or connect to a sensor by COM
+# error, sensor = client.obtain_sensor_by_name("WindowsDevice", "//./COM25", 921600)
+
+# or connect to a Bluetooth sensor (LPMS-B2)
+# error, sensor = client.obtain_sensor_by_name("Bluetooth", "00:04:3E:53:E9:9F", 115200)
 
 if not error == openzen.ZenSensorInitError.NoError:
     print ("Error connecting to sensor")
@@ -67,6 +84,48 @@ if not error == openzen.ZenError.NoError:
     sys.exit(1)
 
 print ("Sensor is streaming data: {}".format(is_streaming))
+
+print("\n>> Set and get IMU settings")
+# test to print imu ID
+error = imu.set_int32_property(openzen.ZenImuProperty.Id, 66)
+error, imu_id = imu.get_int32_property(openzen.ZenImuProperty.Id)
+print("IMU ID: {}".format(imu_id))
+
+# test to set freq
+error = imu.set_int32_property(openzen.ZenImuProperty.SamplingRate, 100)
+error, freq = imu.get_int32_property(openzen.ZenImuProperty.SamplingRate)
+print("Sampling rate: {}".format(freq))
+
+# # test CAN settings
+# print("\n>> Set and get CAN settings")
+# # "CanChannelMode"
+# error = imu.set_int32_property(openzen.ZenImuProperty.CanChannelMode, 1)
+# error, channelMode = imu.get_int32_property(openzen.ZenImuProperty.CanChannelMode)
+# print("CanChannelMode: {}".format(channelMode))
+# #"CanPointMode"
+# error = imu.set_int32_property(openzen.ZenImuProperty.CanPointMode, 1)
+# error, CanPointMode = imu.get_int32_property(openzen.ZenImuProperty.CanPointMode)
+# print("CanPointMode: {}".format(CanPointMode))
+# #"CanStartId"
+# error = imu.set_int32_property(openzen.ZenImuProperty.CanStartId, 0)
+# error, CanStartId = imu.get_int32_property(openzen.ZenImuProperty.CanStartId)
+# print("CanStartId: {}".format(CanStartId))
+# #"CanBaudrate"
+# error = imu.set_int32_property(openzen.ZenImuProperty.CanBaudrate, 125)
+# error, CanBaudrate = imu.get_int32_property(openzen.ZenImuProperty.CanBaudrate)
+# print("CanBaudrate: {}".format(CanBaudrate))
+# #"CanMapping"
+# error = imu.set_array_property_int32(openzen.ZenImuProperty.CanMapping, [3, 5, 6, 19, 20, 21, 28, 29, 30, 38, 39, 40, 34, 35, 36, 37])
+# error, CanMapping = imu.get_array_property_int32(openzen.ZenImuProperty.CanMapping)
+# print("CanMapping: {}".format(CanMapping))
+# #"CanHeartbeat"
+# error = imu.set_int32_property(openzen.ZenImuProperty.CanHeartbeat, 5)
+# error, CanHeartbeat = imu.get_int32_property(openzen.ZenImuProperty.CanHeartbeat)
+# print("CanHeartbeat: {}".format(CanHeartbeat))
+
+print()
+
+
 
 ## load the alignment matrix from the sensor
 ## some sensors don't support this (for example IG1, BE1)
@@ -102,11 +161,14 @@ while True:
         zenEvent.component.handle == imu.component.handle:
 
         imu_data = zenEvent.data.imu_data
-        print ("A: {} m/s^2".format(imu_data.a))
-        print ("G: {} degree/s".format(imu_data.g))
+        print ("ts: {} s".format(imu_data.timestamp))
+        print ("A: {} g".format(imu_data.a))
+        print ("G1: {} degree/s".format(imu_data.g1))   # depending on sensor, gyro data is outputted to g1, g2 or both
+        print ("G2: {} degree/s".format(imu_data.g2))   # read more on https://lpresearch.bitbucket.io/openzen/latest/getting_started.html#id1
+        print ("B: {} microT".format(imu_data.b))
 
     runSome = runSome + 1
-    if runSome > 50:
+    if runSome > 5:
         break
 
 print ("Streaming of sensor data complete")