+++ openzen/examples/VisualStudio2017-OpenZenExample/OpenZenExample/OpenZenExample.cpp
@@ -77,9 +77,9 @@ void pollLoop(std::reference_wrapper<ZenClient> client)
                         std::cout << "> Acceleration: \t x = " << event.data.imuData.a[0]
                             << "\t y = " << event.data.imuData.a[1]
                             << "\t z = " << event.data.imuData.a[2] << std::endl;
-                        std::cout << "> Gyro: \t\t x = " << event.data.imuData.g[0]
-                            << "\t y = " << event.data.imuData.g[1]
-                            << "\t z = " << event.data.imuData.g[2] << std::endl;
+                        std::cout << "> Gyro: \t\t x = " << event.data.imuData.g1[0]
+                            << "\t y = " << event.data.imuData.g1[1]
+                            << "\t z = " << event.data.imuData.g1[2] << std::endl;
                     }
                     break;
                 }